+++ ChangeLog.md
@@ -1,4 +1,7 @@
-﻿#### 2015-12-25 20:59 修正 wkeBrowser 中拦截 _blank 导致无法打开链接的问题。
+﻿#### 2015-12-25 23:44 实现几个底层回调，使 window.close、window.blur、window.focus 生效。
+* 实现 ChromeClient::unfocus、ChromeClient::focus、ChromeClient::closeWindowSoon，使 window.close、window.blur、window.focus 生效。
+
+#### 2015-12-25 20:59 修正 wkeBrowser 中拦截 _blank 导致无法打开链接的问题。
 * 修正 wkeBrowser 中拦截 _blank 导致无法打开链接的问题。
 
 #### 2015-12-24 20:27 临时解决 wke 的 vs2010 工程 Debug 模式链接失败的问题。